Class Selection
Changed the behavior of the class selection to auto-schedule a class at full capacity by default. If it cannot find a non-full class it will give unscheduled.

A quirk: if there are more than one classes running at the same timeslot and the one listed first is full but another is free then it will slot you the free one but then if you click on the card it will straight away update it with the course that is full because it's the 'default' course at that timeslot.