drag cancel on arrow click
Made drag not perform if the button in being onDowned.
Also added few tweaks:
- arrows white irrespective of dark mode
- button circle surrounding the arrow is smaller
And made it so that arrows will only appear on classes that have multiple ones to scroll through, and only the first period if a class has multiple periods (e.g. only on the top card of tut-labs).