stopped drag disrupting selected class
If you have a selected class at a certain time period that is not the default class for that time period then if you simply initiated drag by clicking on the class-card (easy to trigger) than it would instantly revert the chosen class back to the default for that timeslot. Now it doesn't.
I.e. Before updating classes it checks that the new one it's updating to is different to the current one.