Feature/ntgl 256 class selection (#210)
* Arrows on class cards

Arrow buttons on either side of the class card.

* Buttons to change class

Buttons to the sides of class info on the class cards, when pressed, will toggle classes to change from left and right on that time slot.

* drag cancel on arrow click

Made drag not perform if the button in being onDowned.
Also added few tweaks:
- arrows white irrespective of dark mode
- button circle surrounding the arrow is smaller
And made it so that arrows will only appear on classes that have multiple ones to scroll through, and only the first period if a class has multiple periods (e.g. only on the top card of tut-labs).

* stopped drag disrupting selected class

If you have a selected class at a certain time period that is not the default class for that time period then if you simply initiated drag by clicking on the class-card (easy to trigger) than it would instantly revert the chosen class back to the default for that timeslot. Now it doesn't.
I.e. Before updating classes it checks that the new one it's updating to is different to the current one.

* fixed double import

* fixed location calculation

 avoids over calculating / duplication of locations.

* Fixed style

Co-authored-by: Mun Joon Teo <teomunjoon101@gmail.com>